requirements.txt
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
librosa==0.10.2
soundfile==0.12.1
pydub==0.25.1
ffmpeg-python==0.2.0
faiss-cpu==1.8.0
tqdm==4.66.4
certifi>=2024.2.2
laion-clap==1.1.7